#69: Use CommonStoryDecorator instead of global provider
Global provider had life-cycle issues with redux/theming (see log)
